+++ src/main/java/edu/example/tripnote/domain/trip/TourLocDTO.java
@@ -8,7 +8,7 @@ import lombok.ToString;
 @Setter
 @ToString
 public class TourLocDTO {
-	private int id;
+	private Long id;
 	private String code;
 	private String tourLocName;
 	private int tourOrder;
@@ -17,7 +17,7 @@ public class TourLocDTO {
 	private int tourNth;
 	private int timeTaken;
 	private String imgSrc;
-	private int courseId;
+	private Long courseId;
 	private String typeName;
 	private String noteContent;
 	private String mapx;
